chore: logic to hooks
